+++ test/integration/bigip/misc/shareAddresses.js
@@ -436,7 +436,13 @@ describe('shareAddresses', function () {
                 })
                 // Virtual-addresses should be removed from /Tenant once associated apps are removed
                 .then(() => deleteDeclaration(undefined, { path: `${perAppPath}/appOne?async=true`, sendDelete: true }))
+                .then((response) => {
+                    assert.strictEqual(response.results[0].code, 200);
+                })
                 .then(() => deleteDeclaration(undefined, { path: `${perAppPath}/appTwo?async=true`, sendDelete: true }))
+                .then((response) => {
+                    assert.strictEqual(response.results[0].code, 200);
+                })
                 .then(() => getPath('/mgmt/tm/ltm/virtual-address?$filter=partition%20eq%20Tenant'))
                 .then((response) => {
                     assert.strictEqual((response.items || []).length, 0, 'virtual-addresses should be removed from Tenant');